refactor: split review api code across files
1. New Package Structure:
  - pkg/models: Contains shared structs like Repo, PR, Issue, etc.
  - pkg/store: Contains Redis client initialization and mock data population.
  - pkg/k8s: Contains Kubernetes client management and helper functions (e.g., BootstrapNamespace, UpdateSecret, ScaledownSandbox).
  - pkg/auth: Contains authentication logic (GitHub OAuth, session middleware).
  - pkg/api: Contains the Gin server setup (Server struct), route registration, handlers (PRs, Issues, Dev Sandbox, Repo, Settings), and utility functions.
2. Code Refactoring:
  - Global variables in main.go (rdb, k8sClient, oauthConf, etc.) have been removed.
  - Dependencies are now injected via the Server, Authenticator, and k8s.Manager structs.
  - Imports have been updated to reflect the new location of the code.
3. Tests:
  - fix_yaml_integers_test.go and main_test.go were moved to pkg/api/ as fix_yaml_integers_test.go and suggestions_test.go respectively, and verified to pass.
