+++ repo-agent/images/review-api/Dockerfile
@@ -9,8 +9,8 @@ COPY go.sum go.sum
 # and so that source changes don't invalidate our downloaded layer
 RUN go mod download
 
-COPY review-ui/review-api/ .
-RUN CGO_ENABLED=0 GOOS=linux go build -o /review-api .
+COPY review-ui/review-api/ review-ui/review-api/
+RUN CGO_ENABLED=0 GOOS=linux go build -o /review-api review-ui/review-api/main.go
 
 # Final stage
 FROM gcr.io/distroless/static:nonroot
