+++ repo-agent/review-ui/review-api/main.go
@@ -1,196 +1,31 @@
 package main
 
 import (
-	"bytes"
 	"context"
 	"crypto/rand"
 	"encoding/base64"
-	"fmt"
-	"io"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
-	"sort"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/gin-contrib/sessions"
 	"github.com/gin-contrib/sessions/cookie"
 	"github.com/gin-gonic/gin"
-	redis "github.com/go-redis/redis/v8"
 
-	//"github.com/google/go-github/github"
-	"github.com/google/go-github/v39/github"
-	yaml "go.yaml.in/yaml/v3"
-	"golang.org/x/oauth2"
-	githuboauth "golang.org/x/oauth2/github"
-	corev1 "k8s.io/api/core/v1"
-	rbacv1 "k8s.io/api/rbac/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
+	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/review-ui/review-api/pkg/api"
+	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/review-ui/review-api/pkg/auth"
+	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/review-ui/review-api/pkg/k8s"
+	"github.com/gke-labs/gemini-for-kubernetes-development/repo-agent/review-ui/review-api/pkg/store"
+
 	"k8s.io/client-go/dynamic"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
-	"k8s.io/client-go/util/retry"
-)
-
-var (
-	rdb          *redis.Client
-	k8sClient    dynamic.Interface
-	k8sClientset *kubernetes.Clientset
-	oauthConf    *oauth2.Config
-	oauthState   string
-	allowedUsers []string // List of allowed GitHub user handles
 )
 
 const (
-	sessionName      = "repo-agent-session"
-	userKey          = "ghUser"
-	systemNamespace  = "repo-agent-system"
-	githubSecretName = "github-pat"
-	geminiSecretName = "gemini-vscode-tokens"
-	claudeSecretName = "anthropic-api-key"
-	devContainerCM   = "devcontainer-json"
+	sessionName = "repo-agent-session"
 )
-
-// AgentOutput defines the structure for the agent's YAML output.
-type AgentOutput struct {
-	Note   string                           `yaml:"note"`
-	Review *github.PullRequestReviewRequest `yaml:"review"`
-}
-
-// PR represents a pull request
-type PR struct {
-	ID             string `json:"id"`
-	Title          string `json:"title"`
-	Draft          string `json:"draft,omitempty"`
-	Sandbox        string `json:"sandbox,omitempty"`
-	SandboxReplica string `json:"sandboxReplica,omitempty"`
-	Review         string `json:"review,omitempty"`
-	HTMLURL        string `json:"htmlURL,omitempty"`
-	DiffURL        string `json:"diffURL,omitempty"`
-}
-
-// Issue represents a GitHub issue
-type Issue struct {
-	ID             string `json:"id"`
-	Title          string `json:"title"`
-	Draft          string `json:"draft,omitempty"`
-	Sandbox        string `json:"sandbox,omitempty"`
-	SandboxReplica string `json:"sandboxReplica,omitempty"`
-	Comment        string `json:"comment,omitempty"`
-	HTMLURL        string `json:"htmlURL,omitempty"`
-	BranchURL      string `json:"branchURL,omitempty"`
-	PushBranch     bool   `json:"pushBranch"`
-}
-
-// Repo represents a repository with its configuration
-type Repo struct {
-	Name          string         `json:"name"`
-	Namespace     string         `json:"namespace"`
-	URL           string         `json:"url"`
-	Review        *ReviewConfig  `json:"review,omitempty"`
-	IssueHandlers []IssueHandler `json:"issueHandlers,omitempty"`
-	Dev           *DevConfig     `json:"dev,omitempty"`
-}
-
-// ReviewConfig holds configuration for PR reviews
-type ReviewConfig struct {
-	MaxActiveSandboxes int64 `json:"maxActiveSandboxes"`
-}
-
-// IssueHandler holds configuration for an issue handler
-type IssueHandler struct {
-	Name               string `json:"name"`
-	MaxActiveSandboxes int64  `json:"maxActiveSandboxes"`
-	PushBranch         bool   `json:"pushBranch"`
-}
-
-// DevConfig holds configuration for dev sandboxes
-type DevConfig struct {
-	MaxActiveSandboxes int64 `json:"maxActiveSandboxes"`
-}
-
-// DevSandbox represents a dev sandbox
-type DevSandbox struct {
-	Name           string `json:"name"`
-	Sandbox        string `json:"sandbox,omitempty"`
-	SandboxReplica string `json:"sandboxReplica,omitempty"`
-	BranchURL      string `json:"branchURL,omitempty"`
-	Branch         string `json:"branch,omitempty"`
-}
-
-type bodyLogWriter struct {
-	gin.ResponseWriter
-	body *bytes.Buffer
-}
-
-func (w bodyLogWriter) Write(b []byte) (int, error) {
-	w.body.Write(b)
-	return w.ResponseWriter.Write(b)
-}
-
-func RequestLoggerMiddleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		// Read the request body
-		var bodyBytes []byte
-		if c.Request.Body != nil {
-			bodyBytes, _ = io.ReadAll(c.Request.Body)
-			// Restore the io.ReadCloser to its original state for subsequent handlers
-			c.Request.Body = io.NopCloser(bytes.NewBuffer(bodyBytes))
-		}
-
-		log.Printf("Request Method: %s\n", c.Request.Method)
-		log.Printf("Request URL: %s\n", c.Request.URL.String())
-		//log.Printf("Request Headers: %v\n", c.Request.Header)
-		log.Printf("Request Body: %s\n", string(bodyBytes))
-
-		c.Next() // Process the request further
-	}
-}
-
-func ResponseLoggerMiddleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		blw := &bodyLogWriter{body: bytes.NewBufferString(""), ResponseWriter: c.Writer}
-		c.Writer = blw
-
-		c.Next() // Process the request and generate the response
-
-		log.Printf("Response Status: %d\n", c.Writer.Status())
-		log.Printf("Response Headers: %v\n", c.Writer.Header())
-		log.Printf("Response Body: %s\n", blw.body.String())
-	}
-}
-
-func initOAuth() {
-	clientID := os.Getenv("GITHUB_CLIENT_ID")
-	clientSecret := os.Getenv("GITHUB_CLIENT_SECRET")
-
-	if clientID == "" || clientSecret == "" {
-		log.Println("Warning: GITHUB_CLIENT_ID or GITHUB_CLIENT_SECRET not set. OAuth will not work.")
-	}
-
-	oauthConf = &oauth2.Config{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		Scopes:       []string{"read:user", "user:email"},
-		Endpoint:     githuboauth.Endpoint,
-		// For GitHub Apps, permissions are defined on the app itself during installation.
-		// However, for OAuth Apps (legacy), the scopes parameter is used to request specific permissions.
-		// We set default scopes here, but they can be overridden per request.
-	}
-
-	b := make([]byte, 16)
-	if _, err := rand.Read(b); err != nil {
-		log.Fatalf("Failed to generate random OAuth state: %v", err)
-	}
-	oauthState = base64.URLEncoding.EncodeToString(b)
-}
 
 func main() {
 	// Redis client
@@ -198,9 +33,16 @@ func main() {
 	if redisAddr == "" {
 		redisAddr = "localhost:6379"
 	}
-	rdb = redis.NewClient(&redis.Options{
-		Addr: redisAddr,
-	})
+	rdb := store.NewClient(redisAddr)
+
+	// Ping redis to ensure connection
+	_, err := rdb.Ping(context.Background()).Result()
+	if err != nil {
+		log.Fatalf("Failed to connect to Redis: %v", err)
+	}
+
+	// Pre-populate mock data in Redis
+	store.PopulateMockData(context.Background(), rdb)
 
 	// Kubernetes client
 	config, err := rest.InClusterConfig()
@@ -215,18 +57,20 @@ func main() {
 			log.Fatalf("Failed to get local kubeconfig: %v", err)
 		}
 	}
-	k8sClient, err = dynamic.NewForConfig(config)
+	k8sClient, err := dynamic.NewForConfig(config)
 	if err != nil {
 		log.Fatalf("Failed to create kubernetes client: %v", err)
 	}
-	k8sClientset, err = kubernetes.NewForConfig(config)
+	k8sClientset, err := kubernetes.NewForConfig(config)
 	if err != nil {
 		log.Fatalf("Failed to create clientset: %v", err)
 	}
 
-	initOAuth()
+	// K8s Manager
+	k8sManager := k8s.NewManager(k8sClient, k8sClientset, rdb)
 
-	// Initialize the GitHub allowed users list
+	// Allowed Users
+	var allowedUsers []string
 	if allowedUsersStr := os.Getenv("GITHUB_ALLOWED_USERS"); allowedUsersStr != "" {
 		allowedUsers = strings.Split(allowedUsersStr, ",")
 		log.Printf("GitHub authentication restricted to users: %v", allowedUsers)
@@ -234,14 +78,11 @@ func main() {
 		log.Println("No GITHUB_ALLOWED_USERS environment variable set. All GitHub users allowed to authenticate.")
 	}
 
-	// Ping redis to ensure connection
-	_, err = rdb.Ping(context.Background()).Result()
-	if err != nil {
-		log.Fatalf("Failed to connect to Redis: %v", err)
-	}
+	// Authenticator
+	authenticator := auth.NewAuthenticator(k8sManager, allowedUsers)
 
-	// Pre-populate mock data in Redis
-	populateMockData()
+	// API Server
+	server := api.NewServer(k8sManager, authenticator)
 
 	// Gin router
 	router := gin.Default()
@@ -257,2359 +98,11 @@ func main() {
 	store := cookie.NewStore([]byte(sessionSecret))
 	router.Use(sessions.Sessions(sessionName, store))
 
-	// Add middleware to log requests and responses
-	router.Use(RequestLoggerMiddleware())
-	router.Use(ResponseLoggerMiddleware())
-
-	// Public routes: Authentication flows and initial configuration do not require a session.
-	router.GET("/", healthCheckOk)
-	router.GET("/api/", healthCheckOk)
-	router.GET("/api/auth/login", authLogin)
-	router.GET("/api/auth/callback", authCallback)
-	router.GET("/api/auth/status", authStatus)
-	router.POST("/api/auth/logout", authLogout)
-	router.GET("/api/auth/providers", getAuthProviders)
-	router.POST("/api/auth/github-config", updateGithubConfig)
-
-	// Protected routes: The "api" group applies authMiddleware to enforce valid sessions for all downstream endpoints.
-	api := router.Group("/api")
-	api.Use(authMiddleware())
-	{
-		api.GET("/repos", getRepos)
-		api.POST("/repos", createRepoWatch)
-		api.GET("/getRepoWatch", getDefaultRepoWatch)
-		api.GET("/repos/:repo/yaml", getRepoWatchYAML)
-		api.PUT("/repos/:repo", updateRepoWatch)
-		api.DELETE("/repos/:repo", deleteRepoWatch)
-
-		api.GET("/settings", getSettings)
-		api.POST("/settings", updateSettings)
-
-		api.GET("/repo/:repo/prs", getPRs)
-		api.POST("/repo/:repo/prs/:id/draft", saveDraft)
-		api.POST("/repo/:repo/prs/:id/submitreview", submitReview)
-		api.DELETE("/repo/:repo/prs/:id", deletePR)
-		api.GET("/repo/:repo/issues/:handler", getIssues)
-		api.POST("/repo/:repo/issues/:issue_id/handler/:handler/draft", saveIssueDraft)
-		api.POST("/repo/:repo/issues/:issue_id/handler/:handler/submitcomment", submitIssueComment)
-		api.DELETE("/repo/:repo/issues/:issue_id/handler/:handler", deleteIssue)
-		api.POST("/repo/:repo/prs/:id/scaleup", scaleUpPR)
-		api.POST("/repo/:repo/prs/:id/scaledown", scaleDownPR)
-		api.POST("/repo/:repo/issues/:issue_id/handler/:handler/scaleup", scaleUpIssue)
-		api.POST("/repo/:repo/issues/:issue_id/handler/:handler/scaledown", scaleDownIssue)
-		api.GET("/repo/:repo/dev", getDevSandboxes)
-		api.DELETE("/repo/:repo/dev/:name", deleteDevSandbox)
-		api.POST("/repo/:repo/dev/:name/scaleup", scaleUpDevSandbox)
-		api.POST("/repo/:repo/dev/:name/scaledown", scaleDownDevSandbox)
-		api.GET("/proxy", proxy)
-	}
+	// Register Routes
+	server.RegisterRoutes(router)
 
 	err = router.Run(":8080")
 	if err != nil {
 		log.Fatalf("Failed to start router: %v", err)
 	}
-}
-
-// --- Health Check ---
-func healthCheckOk(c *gin.Context) {
-	c.String(http.StatusOK, "OK")
-}
-
-// --- Auth Handlers ---
-
-func isUserAllowed(username string) bool {
-	if len(allowedUsers) == 0 {
-		return true // No restrictions
-	}
-	for _, allowed := range allowedUsers {
-		if strings.EqualFold(username, allowed) {
-			return true
-		}
-	}
-	return false
-}
-
-func authLogin(c *gin.Context) {
-	if oauthConf.ClientID == "" {
-		c.String(http.StatusInternalServerError, "GitHub OAuth is not configured. Please set GITHUB_CLIENT_ID and GITHUB_CLIENT_SECRET in the github-token secret.")
-		return
-	}
-	scheme := "http"
-	if c.Request.TLS != nil || c.Request.Header.Get("X-Forwarded-Proto") == "https" {
-		scheme = "https"
-	}
-
-	// Create a local copy of the config to modify scopes per request
-	localConf := *oauthConf
-	localConf.RedirectURL = fmt.Sprintf("%s://%s/api/auth/callback", scheme, c.Request.Host)
-
-	scope := c.Query("scope")
-	if scope == "readwrite" {
-		localConf.Scopes = []string{"repo", "read:user", "user:email"}
-	} else {
-		// Default to read-only (or whatever the default was)
-		localConf.Scopes = []string{"read:user", "user:email"}
-	}
-
-	url := localConf.AuthCodeURL(oauthState, oauth2.AccessTypeOnline)
-	c.Redirect(http.StatusTemporaryRedirect, url)
-}
-
-func authCallback(c *gin.Context) {
-	if c.Query("state") != oauthState {
-		c.String(http.StatusBadRequest, "Invalid OAuth state")
-		return
-	}
-	token, err := oauthConf.Exchange(c.Request.Context(), c.Query("code"))
-	if err != nil {
-		log.Printf("OAuth exchange failed: %v", err)
-		c.String(http.StatusInternalServerError, "Authentication failed")
-		return
-	}
-
-	client := github.NewClient(oauthConf.Client(c.Request.Context(), token))
-	user, _, err := client.Users.Get(c.Request.Context(), "")
-	if err != nil {
-		log.Printf("Failed to get GitHub user: %v", err)
-		c.String(http.StatusInternalServerError, "Failed to get user info")
-		return
-	}
-
-	ghUser := strings.ToLower(user.GetLogin())
-
-	// Enforce allowlist for GitHub users
-	if !isUserAllowed(ghUser) {
-		log.Printf("Unauthorized GitHub user attempted to log in: %s", ghUser)
-		c.String(http.StatusForbidden, "Unauthorized GitHub user. Please contact administrator.")
-		return
-	}
-
-	if err := bootstrapNamespace(c.Request.Context(), ghUser); err != nil {
-		log.Printf("Failed to bootstrap namespace %s: %v", ghUser, err)
-	}
-
-	// Update the secret with the user's token and info
-	if err := updateUserSecret(c.Request.Context(), ghUser, token, user); err != nil {
-		log.Printf("Failed to update user secret: %v", err)
-		c.String(http.StatusInternalServerError, "Failed to update user secret")
-		return
-	}
-
-	session := sessions.Default(c)
-	session.Set(userKey, ghUser)
-	if err := session.Save(); err != nil {
-		log.Printf("Failed to save session: %v", err)
-		c.String(http.StatusInternalServerError, "Failed to save session")
-		return
-	}
-	c.Redirect(http.StatusTemporaryRedirect, "/")
-}
-
-func updateUserSecret(ctx context.Context, namespace string, token *oauth2.Token, user *github.User) error {
-	data := map[string][]byte{
-		"pat": []byte(token.AccessToken),
-	}
-	if token.RefreshToken != "" {
-		data["refresh_token"] = []byte(token.RefreshToken)
-	}
-	if !token.Expiry.IsZero() {
-		data["expiry"] = []byte(token.Expiry.Format(time.RFC3339))
-	}
-	if user.Name != nil {
-		data["name"] = []byte(*user.Name)
-	}
-	if user.Email != nil {
-		data["email"] = []byte(*user.Email)
-	}
-	return updateSecret(ctx, namespace, githubSecretName, data)
-}
-
-func authStatus(c *gin.Context) {
-	session := sessions.Default(c)
-	if user := session.Get(userKey); user != nil {
-		c.JSON(http.StatusOK, gin.H{"authenticated": true, "user": user})
-		return
-	}
-	c.JSON(http.StatusUnauthorized, gin.H{"authenticated": false})
-}
-
-func authLogout(c *gin.Context) {
-	session := sessions.Default(c)
-	session.Delete(userKey)
-	if err := session.Save(); err != nil {
-		log.Printf("Failed to save session: %v", err)
-		c.String(http.StatusInternalServerError, "Failed to save session")
-		return
-	}
-	c.Status(http.StatusOK)
-}
-
-func getAuthProviders(c *gin.Context) {
-	configured := oauthConf.ClientID != "" && oauthConf.ClientSecret != ""
-	c.JSON(http.StatusOK, gin.H{"github": configured})
-}
-
-func updateGithubConfig(c *gin.Context) {
-	var payload struct {
-		ClientID     string `json:"client_id"`
-		ClientSecret string `json:"client_secret"`
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if payload.ClientID == "" || payload.ClientSecret == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "client_id and client_secret are required"})
-		return
-	}
-
-	// Update Secret in repo-agent-system
-	// We need to get the existing secret to preserve the PAT
-	secret, err := k8sClientset.CoreV1().Secrets(systemNamespace).Get(c.Request.Context(), githubSecretName, v1.GetOptions{})
-	if err != nil {
-		log.Printf("Failed to get github secret: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get github secret"})
-		return
-	}
-
-	if secret.Data == nil {
-		secret.Data = make(map[string][]byte)
-	}
-	secret.Data["github-client-id"] = []byte(payload.ClientID)
-	secret.Data["github-client-secret"] = []byte(payload.ClientSecret)
-
-	_, err = k8sClientset.CoreV1().Secrets(systemNamespace).Update(c.Request.Context(), secret, v1.UpdateOptions{})
-	if err != nil {
-		log.Printf("Failed to update github secret: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update github secret"})
-		return
-	}
-
-	// Update in-memory config
-	oauthConf.ClientID = payload.ClientID
-	oauthConf.ClientSecret = payload.ClientSecret
-
-	c.Status(http.StatusOK)
-}
-
-func authMiddleware() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		session := sessions.Default(c)
-		userVal := session.Get(userKey)
-
-		// If no user is logged in, default to "default" namespace (guest mode)
-		// The user requested: "no auth" logic that puts the user in the default namespace
-		user := "default"
-		if userVal != nil {
-			user = userVal.(string)
-		}
-
-		// Lazy bootstrap checks if namespace exists, creating it if needed.
-		if err := bootstrapNamespace(c.Request.Context(), user); err != nil {
-			log.Printf("Lazy bootstrap failed for user %s: %v", user, err)
-		}
-
-		c.Set(userKey, user)
-		c.Next()
-	}
-}
-
-// --- Bootstrap ---
-
-func bootstrapNamespace(ctx context.Context, targetNS string) error {
-	_, err := k8sClientset.CoreV1().Namespaces().Get(ctx, targetNS, v1.GetOptions{})
-	if errors.IsNotFound(err) {
-		log.Printf("Creating namespace %s", targetNS)
-		ns := &corev1.Namespace{
-			ObjectMeta: v1.ObjectMeta{
-				Name:   targetNS,
-				Labels: map[string]string{"app.kubernetes.io/managed-by": "repo-agent", "review.gemini.google.com/tenant": targetNS},
-			},
-		}
-		if _, err := k8sClientset.CoreV1().Namespaces().Create(ctx, ns, v1.CreateOptions{}); err != nil {
-			return err
-		}
-	} else if err != nil {
-		return err
-	}
-
-	// Copy default secrets/configs from system namespace if they don't exist in user namespace
-	if err := copySecret(ctx, systemNamespace, githubSecretName, targetNS, githubSecretName); err != nil {
-		log.Printf("Warning: failed to copy default github secret: %v", err)
-	}
-	if err := copySecret(ctx, systemNamespace, geminiSecretName, targetNS, geminiSecretName); err != nil {
-		log.Printf("Warning: failed to copy default gemini secret: %v", err)
-	}
-	if err := copySecret(ctx, systemNamespace, claudeSecretName, targetNS, claudeSecretName); err != nil {
-		log.Printf("Warning: failed to copy default claude secret: %v", err)
-	}
-	if err := copyConfigMap(ctx, systemNamespace, devContainerCM, targetNS, devContainerCM); err != nil {
-		log.Printf("Debug: failed to copy %s: %v", devContainerCM, err)
-	}
-
-	if err := setupServiceAccounts(ctx, targetNS); err != nil {
-		log.Printf("Warning: failed to setup service accounts: %v", err)
-	}
-
-	return nil
-}
-
-func copySecret(ctx context.Context, srcNS, srcName, dstNS, dstName string) error {
-	src, err := k8sClientset.CoreV1().Secrets(srcNS).Get(ctx, srcName, v1.GetOptions{})
-	if err != nil {
-		log.Printf("Error reading secret %s/%s: %v", srcNS, srcName, err)
-		return err
-	}
-	dst := &corev1.Secret{ObjectMeta: v1.ObjectMeta{Name: dstName, Namespace: dstNS}, Data: src.Data, Type: src.Type}
-	_, err = k8sClientset.CoreV1().Secrets(dstNS).Create(ctx, dst, v1.CreateOptions{})
-	return ignoreAlreadyExists(err)
-}
-
-func copyConfigMap(ctx context.Context, srcNS, srcName, dstNS, dstName string) error {
-	src, err := k8sClientset.CoreV1().ConfigMaps(srcNS).Get(ctx, srcName, v1.GetOptions{})
-	if err != nil {
-		return err
-	}
-	dst := &corev1.ConfigMap{ObjectMeta: v1.ObjectMeta{Name: dstName, Namespace: dstNS}, Data: src.Data, BinaryData: src.BinaryData}
-	_, err = k8sClientset.CoreV1().ConfigMaps(dstNS).Create(ctx, dst, v1.CreateOptions{})
-	return ignoreAlreadyExists(err)
-}
-
-func setupServiceAccounts(ctx context.Context, ns string) error {
-	// --- Review Sandbox ---
-	saReview := &corev1.ServiceAccount{ObjectMeta: v1.ObjectMeta{Name: "review-sandbox", Namespace: ns}}
-	_, err := k8sClientset.CoreV1().ServiceAccounts(ns).Create(ctx, saReview, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// Bind to review-sandbox cluster role (base permissions)
-	rbReview := &rbacv1.RoleBinding{
-		ObjectMeta: v1.ObjectMeta{Name: "review-sandbox-binding", Namespace: ns},
-		Subjects:   []rbacv1.Subject{{Kind: "ServiceAccount", Name: "review-sandbox", Namespace: ns}},
-		RoleRef:    rbacv1.RoleRef{Kind: "ClusterRole", Name: "review-sandbox", APIGroup: "rbac.authorization.k8s.io"},
-	}
-	_, err = k8sClientset.RbacV1().RoleBindings(ns).Create(ctx, rbReview, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// Add to review-sandbox cluster role binding (to match make apply-common-for-examples)
-	if err := ensureClusterRoleBindingSubject(ctx, "review-sandbox", rbacv1.Subject{Kind: "ServiceAccount", Name: "review-sandbox", Namespace: ns}); err != nil {
-		log.Printf("Warning: failed to update review-sandbox cluster role binding: %v", err)
-	}
-
-	// Bind to configdir-controller cluster role (needed for init container)
-	rbReviewConfigDir := &rbacv1.RoleBinding{
-		ObjectMeta: v1.ObjectMeta{Name: "review-sandbox-configdir-binding", Namespace: ns},
-		Subjects:   []rbacv1.Subject{{Kind: "ServiceAccount", Name: "review-sandbox", Namespace: ns}},
-		RoleRef:    rbacv1.RoleRef{Kind: "ClusterRole", Name: "configdir-controller", APIGroup: "rbac.authorization.k8s.io"},
-	}
-	_, err = k8sClientset.RbacV1().RoleBindings(ns).Create(ctx, rbReviewConfigDir, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// --- Dev Sandbox ---
-	saDev := &corev1.ServiceAccount{ObjectMeta: v1.ObjectMeta{Name: "dev-sandbox", Namespace: ns}}
-	_, err = k8sClientset.CoreV1().ServiceAccounts(ns).Create(ctx, saDev, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// Bind to dev-sandbox cluster role (base permissions)
-	rbDev := &rbacv1.RoleBinding{
-		ObjectMeta: v1.ObjectMeta{Name: "dev-sandbox-binding", Namespace: ns},
-		Subjects:   []rbacv1.Subject{{Kind: "ServiceAccount", Name: "dev-sandbox", Namespace: ns}},
-		RoleRef:    rbacv1.RoleRef{Kind: "ClusterRole", Name: "dev-sandbox", APIGroup: "rbac.authorization.k8s.io"},
-	}
-	_, err = k8sClientset.RbacV1().RoleBindings(ns).Create(ctx, rbDev, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// Add to dev-sandbox cluster role binding (to match make apply-common-for-examples)
-	if err := ensureClusterRoleBindingSubject(ctx, "dev-sandbox", rbacv1.Subject{Kind: "ServiceAccount", Name: "dev-sandbox", Namespace: ns}); err != nil {
-		log.Printf("Warning: failed to update dev-sandbox cluster role binding: %v", err)
-	}
-
-	// Bind to configdir-controller cluster role (needed for init container)
-	rbDevConfigDir := &rbacv1.RoleBinding{
-		ObjectMeta: v1.ObjectMeta{Name: "dev-sandbox-configdir-binding", Namespace: ns},
-		Subjects:   []rbacv1.Subject{{Kind: "ServiceAccount", Name: "dev-sandbox", Namespace: ns}},
-		RoleRef:    rbacv1.RoleRef{Kind: "ClusterRole", Name: "configdir-controller", APIGroup: "rbac.authorization.k8s.io"},
-	}
-	_, err = k8sClientset.RbacV1().RoleBindings(ns).Create(ctx, rbDevConfigDir, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// --- Issue Sandbox ---
-	saIssue := &corev1.ServiceAccount{ObjectMeta: v1.ObjectMeta{Name: "issue-sandbox", Namespace: ns}}
-	_, err = k8sClientset.CoreV1().ServiceAccounts(ns).Create(ctx, saIssue, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// Bind to issue-sandbox cluster role (base permissions)
-	rbIssue := &rbacv1.RoleBinding{
-		ObjectMeta: v1.ObjectMeta{Name: "issue-sandbox-binding", Namespace: ns},
-		Subjects:   []rbacv1.Subject{{Kind: "ServiceAccount", Name: "issue-sandbox", Namespace: ns}},
-		RoleRef:    rbacv1.RoleRef{Kind: "ClusterRole", Name: "issue-sandbox", APIGroup: "rbac.authorization.k8s.io"},
-	}
-	_, err = k8sClientset.RbacV1().RoleBindings(ns).Create(ctx, rbIssue, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	// Add to issue-sandbox cluster role binding (to match make apply-common-for-examples)
-	if err := ensureClusterRoleBindingSubject(ctx, "issue-sandbox", rbacv1.Subject{Kind: "ServiceAccount", Name: "issue-sandbox", Namespace: ns}); err != nil {
-		log.Printf("Warning: failed to update issue-sandbox cluster role binding: %v", err)
-	}
-
-	// Bind to configdir-controller cluster role (needed for init container)
-	rbIssueConfigDir := &rbacv1.RoleBinding{
-		ObjectMeta: v1.ObjectMeta{Name: "issue-sandbox-configdir-binding", Namespace: ns},
-		Subjects:   []rbacv1.Subject{{Kind: "ServiceAccount", Name: "issue-sandbox", Namespace: ns}},
-		RoleRef:    rbacv1.RoleRef{Kind: "ClusterRole", Name: "configdir-controller", APIGroup: "rbac.authorization.k8s.io"},
-	}
-	_, err = k8sClientset.RbacV1().RoleBindings(ns).Create(ctx, rbIssueConfigDir, v1.CreateOptions{})
-	if err != nil && !errors.IsAlreadyExists(err) {
-		return err
-	}
-
-	return nil
-}
-
-func ensureClusterRoleBindingSubject(ctx context.Context, bindingName string, subject rbacv1.Subject) error {
-	return retry.RetryOnConflict(retry.DefaultRetry, func() error {
-		binding, err := k8sClientset.RbacV1().ClusterRoleBindings().Get(ctx, bindingName, v1.GetOptions{})
-		if err != nil {
-			return err
-		}
-		for _, s := range binding.Subjects {
-			if s.Kind == subject.Kind && s.Name == subject.Name && s.Namespace == subject.Namespace {
-				return nil // Already exists
-			}
-		}
-		binding.Subjects = append(binding.Subjects, subject)
-		_, err = k8sClientset.RbacV1().ClusterRoleBindings().Update(ctx, binding, v1.UpdateOptions{})
-		return err
-	})
-}
-
-func ignoreAlreadyExists(err error) error {
-	if errors.IsAlreadyExists(err) {
-		return nil
-	}
-	return err
-}
-
-// --- Settings Handlers ---
-
-func getSettings(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	settings := gin.H{"github_pat_set": false, "gemini_api_key_set": false}
-
-	if s, err := k8sClientset.CoreV1().Secrets(namespace).Get(c.Request.Context(), githubSecretName, v1.GetOptions{}); err == nil {
-		if _, ok := s.Data["pat"]; ok {
-			settings["github_pat_set"] = true
-		}
-	}
-	if s, err := k8sClientset.CoreV1().Secrets(namespace).Get(c.Request.Context(), geminiSecretName, v1.GetOptions{}); err == nil {
-		if _, ok := s.Data["gemini"]; ok {
-			settings["gemini_api_key_set"] = true
-		}
-	}
-	c.JSON(http.StatusOK, settings)
-}
-
-func updateSettings(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	var payload struct {
-		GithubPAT    string `json:"github_pat"`
-		GeminiAPIKey string `json:"gemini_api_key"`
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if payload.GithubPAT != "" {
-		err := updateSecret(c.Request.Context(), namespace, githubSecretName, map[string][]byte{"pat": []byte(payload.GithubPAT)})
-		if err != nil {
-			log.Printf("Failed to update GitHub PAT: %v", err)
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update GitHub PAT"})
-			return
-		}
-	}
-
-	if payload.GeminiAPIKey != "" {
-		err := updateSecret(c.Request.Context(), namespace, geminiSecretName, map[string][]byte{"gemini": []byte(payload.GeminiAPIKey)})
-		if err != nil {
-			log.Printf("Failed to update Gemini API Key: %v", err)
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update Gemini API Key"})
-			return
-		}
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func updateSecret(ctx context.Context, namespace, name string, data map[string][]byte) error {
-	secret, err := k8sClientset.CoreV1().Secrets(namespace).Get(ctx, name, v1.GetOptions{})
-	if errors.IsNotFound(err) {
-		secret = &corev1.Secret{
-			ObjectMeta: v1.ObjectMeta{Name: name, Namespace: namespace},
-			Data:       data,
-		}
-		_, err = k8sClientset.CoreV1().Secrets(namespace).Create(ctx, secret, v1.CreateOptions{})
-		return err
-	} else if err != nil {
-		return err
-	}
-
-	if secret.Data == nil {
-		secret.Data = make(map[string][]byte)
-	}
-	for k, v := range data {
-		secret.Data[k] = v
-	}
-	_, err = k8sClientset.CoreV1().Secrets(namespace).Update(ctx, secret, v1.UpdateOptions{})
-	return err
-}
-
-// --- Repo Management Handlers ---
-
-func createRepoWatch(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	var payload struct {
-		YAML string `json:"yaml"`
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	gvr := schema.GroupVersionResource{
-		Group:    "review.gemini.google.com",
-		Version:  "v1alpha1",
-		Resource: "repowatches",
-	}
-
-	if payload.YAML == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "YAML content is required"})
-		return
-	}
-
-	// Create from YAML
-	var unstructuredObj map[string]interface{}
-	if err := yaml.Unmarshal([]byte(payload.YAML), &unstructuredObj); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid YAML content"})
-		return
-	}
-	unstructuredObj = fixYAMLIntegers(unstructuredObj).(map[string]interface{})
-	repoWatch := &unstructured.Unstructured{Object: unstructuredObj}
-
-	// Enforce namespace
-	repoWatch.SetNamespace(namespace)
-
-	// Auto-populate labels if missing
-	labels, found, _ := unstructured.NestedStringSlice(repoWatch.Object, "spec", "review", "labels")
-	if !found || len(labels) == 0 {
-		// Ensure githubSecretName is set for getGitHubToken to work
-		_, found, _ = unstructured.NestedString(repoWatch.Object, "spec", "githubSecretName")
-		if !found {
-			_ = unstructured.SetNestedField(repoWatch.Object, "github-pat", "spec", "githubSecretName")
-		}
-
-		token, tokenErr := getGitHubToken(c.Request.Context(), repoWatch)
-		if tokenErr == nil {
-			ts := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: token})
-			tc := oauth2.NewClient(c.Request.Context(), ts)
-			client := github.NewClient(tc)
-
-			repoURL, _, _ := unstructured.NestedString(repoWatch.Object, "spec", "repoURL")
-			if owner, repoName, urlErr := parseRepoURL(repoURL); urlErr == nil {
-				suggested, suggestErr := getSuggestedLabels(c.Request.Context(), client, owner, repoName)
-				if suggestErr == nil && len(suggested) > 0 {
-					var suggestedInterface []interface{}
-					for _, s := range suggested {
-						var inner []interface{}
-						for _, l := range s {
-							inner = append(inner, l)
-						}
-						suggestedInterface = append(suggestedInterface, inner)
-					}
-					if setErr := unstructured.SetNestedSlice(repoWatch.Object, suggestedInterface, "spec", "review", "labels"); setErr != nil {
-						log.Printf("Failed to set suggested labels: %v", setErr)
-					}
-				} else if suggestErr != nil {
-					log.Printf("Failed to get suggested labels: %v", suggestErr)
-				}
-			}
-		} else {
-			log.Printf("Debug: Could not get token for label suggestion: %v", tokenErr)
-		}
-	}
-
-	_, err := k8sClient.Resource(gvr).Namespace(namespace).Create(c.Request.Context(), repoWatch, v1.CreateOptions{})
-	if err != nil {
-		log.Printf("Failed to create RepoWatch from YAML: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to create RepoWatch: %v", err)})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func getDefaultRepoWatch(c *gin.Context) {
-	defaultRepoWatch := `
-apiVersion: review.gemini.google.com/v1alpha1
-kind: RepoWatch
-metadata:
-  name: change-name
-spec:
-  repoURL: https://github.com/gke-labs/gemini-for-kubernetes-development
-  pollIntervalSeconds: 300
-  githubSecretName: github-pat
-  dev:
-    maxActiveSandboxes: 1
-    maxSandboxes: 3
-    devcontainerConfigRef: devcontainer-json
-    llm:
-      apiKeySecretRef: gemini-vscode-tokens
-      provider: gemini-cli
-  review:
-    preferAssignedToSelf: true
-    reviewShutdownAfterMinutes: 30
-    devcontainerConfigRef: devcontainer-json
-    llm:
-      apiKeySecretRef: gemini-vscode-tokens
-      prompt: >-
-        You are an expert kubernetes developer who is helping with code reviews.
-        Please look at the most recent commit and provide a review feedback.
-        Would you approve it ?
-        Please pay attention to the following:
-        1. Does the fix resolve the original problem.
-        2. Look for linked issues to understand the original problem.
-        3. Are there tests to check the fix.
-      provider: gemini-cli
-    maxActiveSandboxes: 3
-    maxSandboxes: 5
-  issueHandlers:
-    - llm:
-        apiKeySecretRef: gemini-vscode-tokens
-        prompt: >-
-          You are a helpful assistant that triages GitHub issues for a
-          Kubernetes-related open source project.
-          Your task is to categorize incoming issues based on their content and
-          assign appropriate labels.
-          Analyze the issue title and body to determine the most relevant
-          category from the following options:
-          1. bug: Issues that describe unexpected behavior, errors, or
-          malfunctions in the software.
-          2. feature: Suggestions for new features, enhancements, or
-          improvements to existing functionality.
-          2. cleanup: Suggestions for cleaning up code, removing deprecated
-          features, or improving code quality.
-          3. document: Issues related to documentation errors, omissions, or
-          requests for clarification.
-          4. support: Questions or requests for help regarding the use of the
-          software.
-          5. other: Any issue that does not fit into the above categories.
-
-          Start the response with "/kind <Category>" where <Category> is one of
-          bug , feature , document, support, cleanup or other
-          In the next line, provide a concise explanation of your reasoning for
-          the assigned category.
-
-          Issue Title: "{{.Title}}"
-          Issue Body: "{{.Body}}"
-          HTML URL: "{{.HTMLURL}}"
-        provider: gemini-cli
-      issueShutdownAfterMinutes: 30
-      maxActiveSandboxes: 1
-      maxSandboxes: 3
-      name: triage
-`
-	c.JSON(http.StatusOK, gin.H{"yaml": defaultRepoWatch})
-}
-
-func getRepoWatchYAML(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	name := c.Param("repo")
-
-	repoWatch, err := getRepoWatch(c.Request.Context(), namespace, name)
-	if err != nil {
-		log.Printf("Failed to get RepoWatch %s/%s: %v", namespace, name, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get RepoWatch"})
-		return
-	}
-
-	spec, found, err := unstructured.NestedMap(repoWatch.Object, "spec")
-	if err != nil || !found {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get spec from RepoWatch"})
-		return
-	}
-
-	yamlBytes, err := yaml.Marshal(spec)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to marshal spec to YAML"})
-		return
-	}
-
-	c.JSON(http.StatusOK, gin.H{"yaml": string(yamlBytes)})
-}
-
-func updateRepoWatch(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	name := c.Param("repo")
-
-	var payload struct {
-		RepoURL string `json:"repoURL"`
-		AddPR   int    `json:"addPR"`
-		YAML    string `json:"yaml"`
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	if payload.AddPR == 0 && payload.YAML == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "addPR or yaml is required"})
-		return
-	}
-
-	gvr := schema.GroupVersionResource{
-		Group:    "review.gemini.google.com",
-		Version:  "v1alpha1",
-		Resource: "repowatches",
-	}
-
-	// Get existing resource
-	existing, err := k8sClient.Resource(gvr).Namespace(namespace).Get(c.Request.Context(), name, v1.GetOptions{})
-	if err != nil {
-		log.Printf("Failed to get RepoWatch for update: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to get RepoWatch: %v", err)})
-		return
-	}
-
-	if payload.YAML != "" {
-		var spec map[string]interface{}
-		if err := yaml.Unmarshal([]byte(payload.YAML), &spec); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid YAML content"})
-			return
-		}
-
-		spec = fixYAMLIntegers(spec).(map[string]interface{})
-
-		if err := unstructured.SetNestedField(existing.Object, spec, "spec"); err != nil {
-			log.Printf("Failed to set spec: %v", err)
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update object structure"})
-			return
-		}
-
-		// If repoURL changed in YAML, we should update Redis too, but strictly speaking
-		// we should extract it from the new spec.
-		if newURL, found, _ := unstructured.NestedString(existing.Object, "spec", "repoURL"); found {
-			if err := rdb.HSet(c.Request.Context(), fmt.Sprintf("repo:ns:%s:name:%s", namespace, name), "url", newURL).Err(); err != nil {
-				log.Printf("Failed to update repo URL in Redis for %s: %v", name, err)
-			}
-		}
-	}
-
-	// Add PR if provided
-	if payload.AddPR != 0 {
-		pullRequestsSlice, found, err := unstructured.NestedSlice(existing.Object, "spec", "review", "pullRequests")
-		if err != nil {
-			log.Printf("Failed to get pullRequests: %v", err)
-		}
-
-		var pullRequests []int64
-		if found {
-			for _, v := range pullRequestsSlice {
-				if i, ok := v.(int64); ok {
-					pullRequests = append(pullRequests, i)
-				} else if i, ok := v.(int); ok {
-					pullRequests = append(pullRequests, int64(i))
-				}
-			}
-		}
-
-		// Check for duplicates
-		exists := false
-		for _, pr := range pullRequests {
-			if pr == int64(payload.AddPR) {
-				exists = true
-				break
-			}
-		}
-
-		if !exists {
-			pullRequests = append(pullRequests, int64(payload.AddPR))
-			if err := unstructured.SetNestedSlice(existing.Object, convInt64SliceToInterfaceSlice(pullRequests), "spec", "review", "pullRequests"); err != nil {
-				log.Printf("Failed to set pullRequests: %v", err)
-				c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update object structure for pullRequests"})
-				return
-			}
-		}
-	}
-
-	// Apply update
-	_, err = k8sClient.Resource(gvr).Namespace(namespace).Update(c.Request.Context(), existing, v1.UpdateOptions{})
-	if err != nil {
-		log.Printf("Failed to update RepoWatch: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to update RepoWatch: %v", err)})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func convInt64SliceToInterfaceSlice(in []int64) []interface{} {
-	out := make([]interface{}, len(in))
-	for i, v := range in {
-		out[i] = v
-	}
-	return out
-}
-
-func fixYAMLIntegers(in interface{}) interface{} {
-	switch v := in.(type) {
-	case int:
-		return int64(v)
-	case map[string]interface{}:
-		out := make(map[string]interface{}, len(v))
-		for k, val := range v {
-			out[k] = fixYAMLIntegers(val)
-		}
-		return out
-	case []interface{}:
-		out := make([]interface{}, len(v))
-		for i, val := range v {
-			out[i] = fixYAMLIntegers(val)
-		}
-		return out
-	default:
-		return v
-	}
-}
-
-func deleteRepoWatch(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	name := c.Param("repo")
-
-	gvr := schema.GroupVersionResource{
-		Group:    "review.gemini.google.com",
-		Version:  "v1alpha1",
-		Resource: "repowatches",
-	}
-
-	err := k8sClient.Resource(gvr).Namespace(namespace).Delete(c.Request.Context(), name, v1.DeleteOptions{})
-	if err != nil {
-		log.Printf("Failed to delete RepoWatch: %v", err)
-		if !errors.IsNotFound(err) {
-			c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to delete RepoWatch: %v", err)})
-			return
-		}
-	}
-
-	// Also delete from Redis
-	if err := rdb.Del(c.Request.Context(), fmt.Sprintf("repo:ns:%s:name:%s", namespace, name)).Err(); err != nil {
-		log.Printf("Failed to delete repo %s from Redis: %v", name, err)
-		// Don't fail the request if Redis fails, as K8s deletion is the source of truth
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func populateMockData() {
-	ctx := context.Background()
-	mockRepos := []struct {
-		Name string
-		URL  string
-	}{
-		{Name: "redis", URL: "https://github.com/redis/redis"},
-		{Name: "linux", URL: "https://github.com/linux/linux"},
-	}
-
-	mockPRs := map[string][]PR{
-		"redis": {
-			{ID: "123", Title: "Feat: Add awesome feature", Draft: "This is a draft review."},
-			{ID: "124", Title: "Fix: A really bad bug", Sandbox: "redis-pr-124", Review: "LGTM!"},
-		},
-		"linux": {
-			{ID: "1", Title: "Docs: Update README", Sandbox: "linux-pr-1", Draft: "Few spelling mistakes. s/Nort/North/"},
-			{ID: "2", Title: "Refactor: Improve performance"},
-		},
-	}
-
-	for _, repo := range mockRepos {
-		// Store repo URL (Mock data in default namespace)
-		if err := rdb.HSet(ctx, fmt.Sprintf("repo:ns:default:name:%s", repo.Name), "url", repo.URL, "namespace", "default").Err(); err != nil {
-			log.Printf("Failed to set repo URL in Redis: %v", err)
-		}
-
-		// Store PRs for the repo
-		for _, pr := range mockPRs[repo.Name] {
-			prKey := fmt.Sprintf("pr:ns:default:repo:%s:pr:%s", repo.Name, pr.ID)
-			if err := rdb.HSet(ctx, prKey, "title", pr.Title, "draft", pr.Draft, "sandbox", pr.Sandbox, "review", pr.Review).Err(); err != nil {
-				log.Printf("Failed to set PR info in Redis: %v", err)
-			}
-		}
-	}
-}
-
-func getRepos(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	fetchAndPopulateRepos(c.Request.Context(), namespace)
-
-	repos := []Repo{}
-	prefix := fmt.Sprintf("repo:ns:%s:name:", namespace)
-	iter := rdb.Scan(c.Request.Context(), 0, prefix+"*", 0).Iterator()
-	for iter.Next(c.Request.Context()) {
-		key := iter.Val()
-		repoName := key[len(prefix):]
-
-		repoWatch, err := getRepoWatch(c.Request.Context(), namespace, repoName)
-		if err != nil {
-			log.Printf("Failed to get RepoWatch %s/%s: %v", namespace, repoName, err)
-			continue
-		}
-
-		repoURL, found, _ := unstructured.NestedString(repoWatch.Object, "spec", "repoURL")
-		if !found {
-			log.Printf("repoURL not found in RepoWatch CR %s", repoWatch.GetName())
-			continue
-		}
-
-		repo := Repo{
-			Name:      repoName,
-			Namespace: namespace,
-			URL:       repoURL,
-		}
-
-		// Extract review config
-		if maxActiveSandboxes, found, err := unstructured.NestedInt64(repoWatch.Object, "spec", "review", "maxActiveSandboxes"); err == nil && found && maxActiveSandboxes > 0 {
-			repo.Review = &ReviewConfig{MaxActiveSandboxes: maxActiveSandboxes}
-		}
-
-		// Extract dev config
-		if maxActiveSandboxes, found, err := unstructured.NestedInt64(repoWatch.Object, "spec", "dev", "maxActiveSandboxes"); err == nil && found && maxActiveSandboxes > 0 {
-			repo.Dev = &DevConfig{MaxActiveSandboxes: maxActiveSandboxes}
-		}
-
-		// Extract issue handlers
-		if handlers, found, err := unstructured.NestedSlice(repoWatch.Object, "spec", "issueHandlers"); err == nil && found {
-			var issueHandlers []IssueHandler
-			for _, h := range handlers {
-				handlerMap, ok := h.(map[string]interface{})
-				if !ok {
-					continue
-				}
-				name, _ := handlerMap["name"].(string)
-				maxActiveSandboxes, _ := handlerMap["maxActiveSandboxes"].(int64)
-				pushBranch, _ := handlerMap["pushBranch"].(bool)
-
-				if maxActiveSandboxes > 0 {
-					issueHandlers = append(issueHandlers, IssueHandler{
-						Name:               name,
-						MaxActiveSandboxes: maxActiveSandboxes,
-						PushBranch:         pushBranch,
-					})
-				}
-			}
-			repo.IssueHandlers = issueHandlers
-		}
-
-		repos = append(repos, repo)
-	}
-	if err := iter.Err(); err != nil {
-		log.Printf("Error during Redis SCAN: %v", err)
-	}
-
-	c.JSON(http.StatusOK, repos)
-}
-
-func fetchAndPopulateRepos(ctx context.Context, namespace string) {
-	gvr := schema.GroupVersionResource{
-		Group:    "review.gemini.google.com",
-		Version:  "v1alpha1",
-		Resource: "repowatches",
-	}
-	list, err := k8sClient.Resource(gvr).Namespace(namespace).List(context.Background(), v1.ListOptions{})
-	if err != nil {
-		log.Printf("Failed to list RepoWatch CRs: %v. Serving mock data.", err)
-		return
-	}
-
-	for _, item := range list.Items {
-		repoURL, found, err := unstructured.NestedString(item.Object, "spec", "repoURL")
-		if err != nil || !found {
-			log.Printf("repoURL not found in RepoWatch CR %s", item.GetName())
-			continue
-		}
-		// Ensure the URL is in Redis
-		if err := rdb.HSet(ctx, fmt.Sprintf("repo:ns:%s:name:%s", namespace, item.GetName()), "url", repoURL, "namespace", namespace).Err(); err != nil {
-			log.Printf("Failed to cache repo URL for %s: %v", item.GetName(), err)
-		}
-	}
-}
-
-func getPRs(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	fetchAndPopulatePRs(c.Request.Context(), namespace, repo)
-	// SCAN Redis for PRs for repo
-	prs := []PR{}
-	repoPRKeyPrefix := fmt.Sprintf("pr:ns:%s:repo:%s:pr:", namespace, repo)
-	iter := rdb.Scan(c.Request.Context(), 0, repoPRKeyPrefix+"*", 0).Iterator()
-	for iter.Next(c.Request.Context()) {
-		key := iter.Val()
-		prID := key[len(repoPRKeyPrefix):]
-		prData, err := rdb.HGetAll(c.Request.Context(), key).Result()
-		if err != nil {
-			log.Printf("Failed to get PR %s from Redis for repo %s: %v", prID, repo, err)
-			continue
-		}
-		pr := PR{
-			ID:    prID,
-			Title: prData["title"],
-		}
-
-		if _, ok := prData["htmlurl"]; ok {
-			pr.HTMLURL = prData["htmlurl"]
-		}
-		if _, ok := prData["diffurl"]; ok {
-			pr.DiffURL = prData["diffurl"]
-		}
-		if _, ok := prData["draft"]; ok {
-			pr.Draft = prData["draft"]
-		}
-		if _, ok := prData["sandbox"]; ok {
-			pr.Sandbox = prData["sandbox"]
-		}
-		if _, ok := prData["sandboxReplica"]; ok {
-			pr.SandboxReplica = prData["sandboxReplica"]
-		}
-		if _, ok := prData["review"]; ok {
-			pr.Review = prData["review"]
-		}
-		prs = append(prs, pr)
-	}
-	if err := iter.Err(); err != nil {
-		log.Printf("Error during Redis SCAN: %v", err)
-	}
-
-	c.JSON(http.StatusOK, prs)
-}
-
-func fetchAndPopulatePRs(ctx context.Context, namespace, repo string) {
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "reviewsandboxes",
-	}
-	list, err := k8sClient.Resource(gvr).Namespace(namespace).List(context.Background(),
-		v1.ListOptions{
-			LabelSelector: fmt.Sprintf("review.gemini.google.com/repowatch=%s", repo),
-		})
-	if err != nil {
-		log.Printf("Failed to list ReviewSandbox CRs: %v. Serving mock data.", err)
-		return
-	}
-
-	log.Printf("Populating PRs: Found %d reviewsandboxes for Repo: %s", len(list.Items), repo)
-	for _, item := range list.Items {
-		log.Printf("Creating PR entry for ReviewSandbox: %s/%s", item.GetNamespace(), item.GetName())
-		// Get replicas and if it scaled down skip
-		replicas, found, err := unstructured.NestedInt64(item.Object, "spec", "replicas")
-		if err != nil || !found {
-			log.Printf("Replicas (.spec.replicas) not found in ReviewSandbox  %s", item.GetName())
-			continue
-		}
-
-		prID, found, err := unstructured.NestedString(item.Object, "spec", "source", "pr")
-		if err != nil || !found {
-			log.Printf("PR ID (.spec.source.pr) not found in ReviewSandbox  %s", item.GetName())
-			continue
-		}
-
-		title, found, err := unstructured.NestedString(item.Object, "spec", "source", "title")
-		if err != nil || !found {
-			log.Printf("Title (.spec.source.title) not found in ReviewSandbox  %s", item.GetName())
-			continue
-		}
-		htmlurl, found, err := unstructured.NestedString(item.Object, "spec", "source", "htmlURL")
-		if err != nil || !found {
-			log.Printf("Title (.spec.source.htmlURL) not found in ReviewSandbox  %s", item.GetName())
-		}
-		diffurl, found, err := unstructured.NestedString(item.Object, "spec", "source", "diffURL")
-		if err != nil || !found {
-			log.Printf("diffURL (.spec.source.diffURL) not found in ReviewSandbox  %s", item.GetName())
-		}
-
-		// get draft from annotation[agentDraft]
-		draft := ""
-		annotations := item.GetAnnotations()
-		if annotations == nil {
-			log.Printf("agentDraft (annotations=nil) not found in ReviewSandbox %s", item.GetName())
-		} else if _, ok := annotations["agentDraft"]; !ok {
-			log.Printf("agentDraft (annotations[agentDraft]) not found in ReviewSandbox %s", item.GetName())
-		} else {
-			draft = annotations["agentDraft"]
-		}
-
-		pr := PR{
-			ID:             prID,
-			Title:          title,
-			Sandbox:        item.GetName(),
-			HTMLURL:        htmlurl,
-			DiffURL:        diffurl,
-			SandboxReplica: fmt.Sprintf("%d", replicas),
-		}
-
-		prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-		// Ensure the URL is in Redis
-		if err := rdb.HSet(ctx, prKey,
-			"title", pr.Title,
-			"sandbox", pr.Sandbox,
-			"htmlurl", pr.HTMLURL,
-			"diffurl", pr.DiffURL,
-			"sandboxReplica", pr.SandboxReplica,
-			"draft", draft,
-			"agentDraft", draft,
-		).Err(); err != nil {
-			log.Printf("Failed to cache PR %s for repo %s: %v", pr.ID, repo, err)
-		}
-	}
-}
-
-func saveDraft(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	prID := c.Param("id")
-	var payload struct {
-		Draft string
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-	err := rdb.HSet(c.Request.Context(), prKey, "draft", payload.Draft).Err()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save draft"})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func submitReview(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	prID := c.Param("id")
-	var payload struct {
-		Review string
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx := c.Request.Context()
-	log.Printf("Submitting review for PR %s in repo %s with review: %s", prID, repo, payload.Review)
-
-	// Get draft and agentDraft from Redis
-	prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-	prData, err := rdb.HGetAll(ctx, prKey).Result()
-	if err != nil {
-		log.Printf("Failed to get PR %s from Redis for repo %s: %v", prID, repo, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get PR data from Redis"})
-		return
-	}
-
-	draft := payload.Review
-	agentDraft := prData["agentDraft"]
-	sandboxName := prData["sandbox"]
-
-	// Get RepoWatch to get repoURL and secret ref
-	repoWatch, err := getRepoWatch(ctx, namespace, repo)
-	if err != nil {
-		log.Printf("Failed to get repowatch %s: %v", repo, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get repowatch config"})
-		return
-	}
-
-	if draft != agentDraft {
-		// Store feedback for fine-tuning
-		prompt, _, _ := unstructured.NestedString(repoWatch.Object, "spec", "review", "gemini", "prompt")
-		configdir, _, _ := unstructured.NestedString(repoWatch.Object, "spec", "review", "gemini", "configdirRef")
-		repoURL, _, _ := unstructured.NestedString(repoWatch.Object, "spec", "repoURL")
-		owner, _, _ := parseRepoURL(repoURL)
-
-		hfKey := fmt.Sprintf("hf:review:githubuser:%s:repo:%s:pr:%s", owner, repo, prID)
-		if err := rdb.HSet(ctx, hfKey,
-			"draft", draft,
-			"agentDraft", agentDraft,
-			"prompt", prompt,
-			"configdir", configdir,
-		).Err(); err != nil {
-			log.Printf("Failed to store feedback for PR %s in repo %s: %v", prID, repo, err)
-			// Continue without failing the review submission
-		}
-
-		if sandboxName != "" {
-			if err := updateReviewSandboxUserDraft(ctx, namespace, sandboxName, draft); err != nil {
-				log.Printf("Failed to update reviewsandbox userDraft for PR %s in repo %s: %v", prID, repo, err)
-				// Not failing the request for this, just logging.
-			}
-		}
-	}
-
-	// Get GitHub token from secret
-	token, err := getGitHubToken(ctx, repoWatch)
-	if err != nil {
-		log.Printf("Failed to get github token for repo %s: %v", repo, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get github token"})
-		return
-	}
-
-	// Create GitHub client
-	ts := oauth2.StaticTokenSource(
-		&oauth2.Token{AccessToken: token},
-	)
-	tc := oauth2.NewClient(ctx, ts)
-	client := github.NewClient(tc)
-
-	// Parse repo URL
-	repoURL, found, err := unstructured.NestedString(repoWatch.Object, "spec", "repoURL")
-	if err != nil || !found {
-		log.Printf("repoURL not found in RepoWatch CR %s", repoWatch.GetName())
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "repoURL not found in RepoWatch CR"})
-		return
-	}
-	owner, repoName, err := parseRepoURL(repoURL)
-	if err != nil {
-		log.Printf("Failed to parse repo url %s: %v", repoURL, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to parse repo url"})
-		return
-	}
-
-	// Get PR number
-	prNumber, err := strconv.Atoi(prID)
-	if err != nil {
-		log.Printf("Failed to parse prID %s: %v", prID, err)
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid pr id"})
-		return
-	}
-
-	// https://docs.github.com/en/rest/pulls/reviews?apiVersion=2022-11-28#create-a-review-for-a-pull-request
-	// Try Unmarshalling the yaml review payload into PullRequestReviewRequest
-	agentOutput := &AgentOutput{}
-	reviewRequest := &github.PullRequestReviewRequest{}
-	err = yaml.Unmarshal([]byte(payload.Review), &agentOutput)
-	if err != nil {
-		log.Printf("Failed to unmarshal review payload: %v", err)
-		reviewRequest.Body = github.String(payload.Review)
-	} else {
-		reviewRequest = agentOutput.Review
-	}
-
-	// Not setting event sets it as a draft
-	reviewRequest.Event = nil
-
-	log.Printf("reviewRequest being created: %v", reviewRequest)
-	review, resp, err := client.PullRequests.CreateReview(ctx, owner, repoName, prNumber, reviewRequest)
-	if err != nil {
-		log.Printf("response: %v", resp)
-		log.Printf("Failed to create review on PR %d: %v", prNumber, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create review on github", "details": err.Error()})
-		return
-	}
-	log.Printf("review created: %v", review)
-	// Set review in Redis
-	err = rdb.HSet(c.Request.Context(), prKey, "review", payload.Review).Err()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save review", "details": err.Error()})
-		return
-	}
-
-	// Delete draft from Redis
-	err = rdb.HSet(c.Request.Context(), prKey, "draft", "").Err()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to clear draft", "details": err.Error()})
-		return
-	}
-
-	// scale down sandbox
-	err = scaledownSandbox(ctx, namespace, repo, prID)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scaledown Sandbox after review submission", "details": err.Error()})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func deletePR(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	prID := c.Param("id")
-	ctx := c.Request.Context()
-
-	if err := scaledownSandbox(ctx, namespace, repo, prID); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete sandbox", "details": err.Error()})
-		return
-	}
-
-	// Clean up Redis keys
-	prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-	if err := rdb.HDel(c.Request.Context(), prKey, "review", "draft", "sandbox", "htmlurl", "title").Err(); err != nil {
-		log.Printf("Failed to HDEL PR data from Redis: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to HDEL PR data from Redis"})
-		return
-	}
-	if err := rdb.Del(c.Request.Context(), prKey).Err(); err != nil {
-		log.Printf("Failed to DEL PR data from Redis: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to DEL PR data from Redis"})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-//nolint:unused
-func deleteSandbox(ctx context.Context, namespace, repo, prID string) error {
-	prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-	sandboxName, err := rdb.HGet(ctx, prKey, "sandbox").Result()
-	if err == redis.Nil {
-		// If sandbox is not in Redis, we can assume it's already deleted or never existed.
-		log.Printf("Sandbox for repo %s, PR %s not found in Redis. Assuming it's already deleted.", repo, prID)
-		return nil
-	} else if err != nil {
-		return fmt.Errorf("failed to get sandbox name from Redis: %w", err)
-	}
-
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "reviewsandboxes",
-	}
-	log.Printf("Deleting sandbox %s", sandboxName)
-	err = k8sClient.Resource(gvr).Namespace(namespace).Delete(ctx, sandboxName, v1.DeleteOptions{})
-	if err != nil {
-		// We can choose to not return an error if it's already gone.
-		return fmt.Errorf("failed to delete sandbox: %w", err)
-	}
-	return nil
-}
-
-func scaledownSandbox(ctx context.Context, namespace, repo, prID string) error {
-	prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-	sandboxName, err := rdb.HGet(ctx, prKey, "sandbox").Result()
-	if err != nil && err != redis.Nil {
-		return fmt.Errorf("failed to get sandbox name from Redis: %w", err)
-	}
-	if sandboxName == "" {
-		sandboxName = fmt.Sprintf("%s-pr-%s", repo, prID)
-	}
-
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "reviewsandboxes",
-	}
-	log.Printf("Scaling down sandbox %s", sandboxName)
-
-	// Set .spec.replicas to 0 and apply the sandbox object
-	sandbox := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "custom.agents.x-k8s.io/v1alpha1",
-			"kind":       "ReviewSandbox",
-			"metadata": map[string]interface{}{
-				"name":      sandboxName,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"replicas": int64(0),
-			},
-		},
-	}
-
-	_, err = k8sClient.Resource(gvr).Namespace(namespace).Apply(ctx, sandboxName,
-		sandbox, v1.ApplyOptions{FieldManager: "review-ui", Force: true})
-	if err != nil {
-		// We can choose to not return an error if it's already gone.
-		return fmt.Errorf("failed to scaledown sandbox: %w", err)
-	}
-	return nil
-}
-
-func updateReviewSandboxUserDraft(ctx context.Context, namespace, sandboxName, userDraft string) error {
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "reviewsandboxes",
-	}
-
-	// Get the existing resource
-	sandbox, err := k8sClient.Resource(gvr).Namespace(namespace).Get(ctx, sandboxName, v1.GetOptions{})
-	if err != nil {
-		return fmt.Errorf("failed to get reviewsandbox %s: %w", sandboxName, err)
-	}
-
-	// update the annotation[agentDraft]
-	if sandbox.GetAnnotations() == nil {
-		sandbox.SetAnnotations(make(map[string]string))
-	}
-	annotations := sandbox.GetAnnotations()
-	annotations["userDraft"] = userDraft
-	sandbox.SetAnnotations(annotations)
-
-	_, err = k8sClient.Resource(gvr).Namespace(namespace).Update(context.TODO(), sandbox, v1.UpdateOptions{})
-	if err != nil {
-		return fmt.Errorf("failed to update reviewsandbox annotation: %w", err)
-	}
-
-	return nil
-}
-
-func getGitHubToken(ctx context.Context, repoWatch *unstructured.Unstructured) (string, error) {
-	secretName, found, err := unstructured.NestedString(repoWatch.Object, "spec", "githubSecretName")
-	if err != nil || !found {
-		return "", fmt.Errorf("githubSecretName not found in repowatch %s", repoWatch.GetName())
-	}
-	secretKey := "pat"
-
-	secretGVR := schema.GroupVersionResource{Version: "v1", Resource: "secrets"}
-	secretUnstructured, err := k8sClient.Resource(secretGVR).Namespace(repoWatch.GetNamespace()).Get(ctx, secretName, v1.GetOptions{})
-	if err != nil {
-		return "", err
-	}
-
-	secretData, found, err := unstructured.NestedStringMap(secretUnstructured.Object, "data")
-	if err != nil || !found {
-		return "", fmt.Errorf("data field not found in secret %s", secretName)
-	}
-
-	tokenBase64, ok := secretData[secretKey]
-	if !ok {
-		return "", fmt.Errorf("key %s not found in secret %s", secretKey, secretName)
-	}
-
-	tokenBytes, err := base64.StdEncoding.DecodeString(tokenBase64)
-	if err != nil {
-		return "", fmt.Errorf("failed to decode token for key %s in secret %s: %w", secretKey, secretName, err)
-	}
-
-	return string(tokenBytes), nil
-}
-
-func parseRepoURL(repoURL string) (string, string, error) {
-	u, err := url.Parse(repoURL)
-	if err != nil {
-		return "", "", err
-	}
-	parts := strings.Split(strings.Trim(u.Path, "/"), "/")
-	if len(parts) != 2 {
-		return "", "", fmt.Errorf("invalid repo url: %s", repoURL)
-	}
-	return parts[0], parts[1], nil
-}
-
-func getRepoWatch(ctx context.Context, namespace, name string) (*unstructured.Unstructured, error) {
-	gvr := schema.GroupVersionResource{
-		Group:    "review.gemini.google.com",
-		Version:  "v1alpha1",
-		Resource: "repowatches",
-	}
-	repoWatch, err := k8sClient.Resource(gvr).Namespace(namespace).Get(ctx, name, v1.GetOptions{})
-	if err != nil {
-		return nil, err
-	}
-	return repoWatch, nil
-}
-
-func getIssues(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	handler := c.Param("handler")
-	fetchAndPopulateIssues(c.Request.Context(), namespace, repo, handler)
-
-	issues := []Issue{}
-	issueKeyPrefix := fmt.Sprintf("issue:ns:%s:repo:%s:handler:%s:issue:*", namespace, repo, handler)
-	iter := rdb.Scan(c.Request.Context(), 0, issueKeyPrefix, 0).Iterator()
-	for iter.Next(c.Request.Context()) {
-		key := iter.Val()
-		// key is issue:repo:REPO:handler:HANDLER:issue:ISSUEID
-		parts := strings.Split(key, ":")
-		if len(parts) != 9 {
-			continue
-		}
-		issueID := parts[8]
-
-		issueData, err := rdb.HGetAll(c.Request.Context(), key).Result()
-		if err != nil {
-			log.Printf("Failed to get Issue %s from Redis for repo %s handler %s: %v", issueID, repo, handler, err)
-			continue
-		}
-		pushBranch, _ := strconv.ParseBool(issueData["pushBranch"])
-		issue := Issue{
-			ID:         issueID,
-			Title:      issueData["title"],
-			PushBranch: pushBranch,
-		}
-
-		if val, ok := issueData["htmlurl"]; ok {
-			issue.HTMLURL = val
-		}
-		if val, ok := issueData["draft"]; ok {
-			issue.Draft = val
-		}
-		if val, ok := issueData["sandbox"]; ok {
-			issue.Sandbox = val
-		}
-		if val, ok := issueData["sandboxReplica"]; ok {
-			issue.SandboxReplica = val
-		}
-		if val, ok := issueData["comment"]; ok {
-			issue.Comment = val
-		}
-		if val, ok := issueData["branchURL"]; ok {
-			issue.BranchURL = val
-		}
-
-		issues = append(issues, issue)
-	}
-	if err := iter.Err(); err != nil {
-		log.Printf("Error during Redis SCAN for issues: %v", err)
-	}
-
-	c.JSON(http.StatusOK, issues)
-}
-
-func fetchAndPopulateIssues(ctx context.Context, namespace, repo, handler string) {
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "issuesandboxes",
-	}
-	list, err := k8sClient.Resource(gvr).Namespace(namespace).List(context.Background(),
-		v1.ListOptions{
-			LabelSelector: fmt.Sprintf("review.gemini.google.com/repowatch=%s,review.gemini.google.com/handler=%s", repo, handler),
-		})
-	if err != nil {
-		log.Printf("Failed to list IssueSandbox CRs: %v.", err)
-		return
-	}
-
-	log.Printf("Populating Issues: Found %d issuesandboxes for Repo: %s Handler: %s", len(list.Items), repo, handler)
-	for _, item := range list.Items {
-		log.Printf("Creating Issue entry for IssueSandbox: %s/%s", item.GetNamespace(), item.GetName())
-		replicas, found, err := unstructured.NestedInt64(item.Object, "spec", "replicas")
-		if err != nil || !found {
-			log.Printf("Replicas (.spec.replicas) not found in IssueSandbox %s", item.GetName())
-			continue
-		}
-
-		issueID, found, err := unstructured.NestedString(item.Object, "spec", "source", "issue")
-		if err != nil || !found {
-			log.Printf("Issue ID (.spec.source.issue) not found in IssueSandbox %s", item.GetName())
-			continue
-		}
-
-		title, found, err := unstructured.NestedString(item.Object, "spec", "source", "title")
-		if err != nil || !found {
-			log.Printf("Title (.spec.source.title) not found in IssueSandbox %s", item.GetName())
-			continue
-		}
-		htmlurl, found, err := unstructured.NestedString(item.Object, "spec", "source", "htmlURL")
-		if err != nil || !found {
-			log.Printf("htmlURL (.spec.source.htmlURL) not found in IssueSandbox %s", item.GetName())
-		}
-
-		// https://github.com/barney-s/kro/tree/issue-753-bugfix
-		// https://github.com/ + .user.login + source.cloneURL repo name + /tree/ + .destination.branch
-		// https://github.com/kubernetes-sigs/kro/compare/main...barney-s:kro:issue-753-bugfix
-		// .source.cloneURL - .git + /compare/main... + .user.login + : + source.cloneURL repo name  + : + .destination.branch
-
-		cloneURL, found, err := unstructured.NestedString(item.Object, "spec", "source", "cloneURL")
-		if err != nil || !found {
-			log.Printf("branchURL (.spec.source.cloneURL) not found in IssueSandbox %s", item.GetName())
-			cloneURL = "https://github.com/noorg/norepo.git"
-		}
-		login, found, err := unstructured.NestedString(item.Object, "spec", "destination", "user", "login")
-		if err != nil || !found {
-			log.Printf("branchURL (.spec.destination.user.login) not found in IssueSandbox %s", item.GetName())
-			login = "nouser"
-		}
-		branch, found, err := unstructured.NestedString(item.Object, "spec", "destination", "branch")
-		if err != nil || !found {
-			log.Printf("branchURL (.spec.destination.branch) not found in IssueSandbox %s", item.GetName())
-			branch = "nobranch"
-		}
-
-		repoParts := strings.Split(strings.TrimSuffix(cloneURL, ".git"), "/")
-		repoName := repoParts[len(repoParts)-1]
-
-		branchURL := fmt.Sprintf("https://github.com/%s/%s/tree/%s", login, repoName, branch)
-
-		pushBranch, found, err := unstructured.NestedBool(item.Object, "spec", "destination", "pushEnabled")
-		if err != nil || !found {
-			log.Printf("pushBranch (.spec.source.pushBranch) not found in IssueSandbox %s", item.GetName())
-		}
-
-		// get draft from annotation[agentDraft]
-		draft := ""
-		annotations := item.GetAnnotations()
-		if annotations == nil {
-			log.Printf("agentDraft (annotations=nil) not found in IssueSandbox %s", item.GetName())
-		} else if _, ok := annotations["agentDraft"]; !ok {
-			log.Printf("agentDraft (annotations[agentDraft]) not found in IssueSandbox %s", item.GetName())
-		} else {
-			draft = annotations["agentDraft"]
-		}
-
-		issueKey := fmt.Sprintf("issue:ns:%s:repo:%s:handler:%s:issue:%s", namespace, repo, handler, issueID)
-		if err := rdb.HSet(ctx, issueKey,
-			"title", title,
-			"sandbox", item.GetName(),
-			"htmlurl", htmlurl,
-			"sandboxReplica", fmt.Sprintf("%d", replicas),
-			"branchURL", branchURL,
-			"draft", draft,
-			"agentDraft", draft,
-			"pushBranch", strconv.FormatBool(pushBranch),
-		).Err(); err != nil {
-			log.Printf("Failed to cache Issue %s for repo %s handler %s: %v", issueID, repo, handler, err)
-		}
-	}
-}
-
-func saveIssueDraft(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	issueID := c.Param("issue_id")
-	handler := c.Param("handler")
-	var payload struct {
-		Draft string
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	issueKey := fmt.Sprintf("issue:ns:%s:repo:%s:handler:%s:issue:%s", namespace, repo, handler, issueID)
-	err := rdb.HSet(c.Request.Context(), issueKey, "draft", payload.Draft).Err()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save draft"})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func submitIssueComment(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	issueID := c.Param("issue_id")
-	handler := c.Param("handler")
-	var payload struct {
-		Comment string
-	}
-	if err := c.ShouldBindJSON(&payload); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	ctx := c.Request.Context()
-	log.Printf("Submitting comment for Issue %s in repo %s with comment: %s", issueID, repo, payload.Comment)
-
-	issueKey := fmt.Sprintf("issue:ns:%s:repo:%s:handler:%s:issue:%s", namespace, repo, handler, issueID)
-	issueData, err := rdb.HGetAll(ctx, issueKey).Result()
-	if err != nil {
-		log.Printf("Failed to get Issue %s from Redis for repo %s handler %s: %v", issueID, repo, handler, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get Issue data from Redis"})
-		return
-	}
-
-	draft := payload.Comment
-	agentDraft := issueData["agentDraft"]
-
-	repoWatch, err := getRepoWatch(ctx, namespace, repo)
-	if err != nil {
-		log.Printf("Failed to get repowatch %s: %v", repo, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get repowatch config"})
-		return
-	}
-
-	if draft != agentDraft {
-		// Store feedback for fine-tuning
-		var prompt, configdir string
-		if handlers, found, err := unstructured.NestedSlice(repoWatch.Object, "spec", "issueHandlers"); err == nil && found {
-			for _, h := range handlers {
-				handlerMap, ok := h.(map[string]interface{})
-				if !ok {
-					continue
-				}
-				name, _ := handlerMap["name"].(string)
-				if name == handler {
-					gemini, ok := handlerMap["gemini"].(map[string]interface{})
-					if ok {
-						prompt, _ = gemini["prompt"].(string)
-						configdir, _ = gemini["configdirRef"].(string)
-					}
-					break
-				}
-			}
-		}
-
-		repoURL, _, _ := unstructured.NestedString(repoWatch.Object, "spec", "repoURL")
-		owner, _, _ := parseRepoURL(repoURL)
-
-		hfKey := fmt.Sprintf("hf:issue:githubuser:%s:repo:%s:handler:%s:pr:%s", owner, repo, handler, issueID)
-		if err := rdb.HSet(ctx, hfKey,
-			"draft", draft,
-			"agentDraft", agentDraft,
-			"prompt", prompt,
-			"configdirname", configdir,
-		).Err(); err != nil {
-			log.Printf("Failed to store feedback for Issue %s in repo %s: %v", issueID, repo, err)
-			// Continue without failing the comment submission
-		}
-	}
-
-	token, err := getGitHubToken(ctx, repoWatch)
-	if err != nil {
-		log.Printf("Failed to get github token for repo %s: %v", repo, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to get github token"})
-		return
-	}
-
-	ts := oauth2.StaticTokenSource(&oauth2.Token{AccessToken: token})
-	tc := oauth2.NewClient(ctx, ts)
-	client := github.NewClient(tc)
-
-	repoURL, found, err := unstructured.NestedString(repoWatch.Object, "spec", "repoURL")
-	if err != nil || !found {
-		log.Printf("repoURL not found in RepoWatch CR %s", repoWatch.GetName())
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "repoURL not found in RepoWatch CR"})
-		return
-	}
-	owner, repoName, err := parseRepoURL(repoURL)
-	if err != nil {
-		log.Printf("Failed to parse repo url %s: %v", repoURL, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to parse repo url"})
-		return
-	}
-
-	issueNumber, err := strconv.Atoi(issueID)
-	if err != nil {
-		log.Printf("Failed to parse issueID %s: %v", issueID, err)
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid issue id"})
-		return
-	}
-
-	comment := &github.IssueComment{Body: &payload.Comment}
-	_, _, err = client.Issues.CreateComment(ctx, owner, repoName, issueNumber, comment)
-	if err != nil {
-		log.Printf("Failed to create comment on Issue %d: %v", issueNumber, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create comment on github"})
-		return
-	}
-
-	err = rdb.HSet(c.Request.Context(), issueKey, "comment", payload.Comment).Err()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to save comment", "details": err.Error()})
-		return
-	}
-
-	err = rdb.HSet(c.Request.Context(), issueKey, "draft", "").Err()
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to clear draft", "details": err.Error()})
-		return
-	}
-
-	err = scaledownIssueSandbox(ctx, namespace, repo, issueID, handler)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scaledown Sandbox after comment submission", "details": err.Error()})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func scaledownIssueSandbox(ctx context.Context, namespace, repo, issueID, handler string) error {
-	sandboxName := fmt.Sprintf("%s-issue-%s-%s", repo, issueID, handler)
-
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "issuesandboxes",
-	}
-	log.Printf("Scaling down issue sandbox %s", sandboxName)
-
-	sandbox := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "custom.agents.x-k8s.io/v1alpha1",
-			"kind":       "IssueSandbox",
-			"metadata": map[string]interface{}{
-				"name":      sandboxName,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"replicas": int64(0),
-			},
-		},
-	}
-
-	_, err := k8sClient.Resource(gvr).Namespace(namespace).Apply(ctx, sandboxName,
-		sandbox, v1.ApplyOptions{FieldManager: "review-ui", Force: true})
-	if err != nil {
-		return fmt.Errorf("failed to scaledown issue sandbox: %w", err)
-	}
-	return nil
-}
-
-func deleteIssue(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	issueID := c.Param("issue_id")
-	handler := c.Param("handler")
-	ctx := c.Request.Context()
-
-	if err := scaledownIssueSandbox(ctx, namespace, repo, issueID, handler); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete sandbox", "details": err.Error()})
-		return
-	}
-
-	issueKey := fmt.Sprintf("issue:ns:%s:repo:%s:handler:%s:issue:%s", namespace, repo, handler, issueID)
-	if err := rdb.Del(c.Request.Context(), issueKey).Err(); err != nil {
-		log.Printf("Failed to DEL Issue data from Redis: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to DEL Issue data from Redis"})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func getDevSandboxes(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	fetchAndPopulateDevSandboxes(c.Request.Context(), namespace, repo)
-
-	sandboxes := []DevSandbox{}
-	// prefix: dev:ns:NAMESPACE:repo:REPO:dev:NAME
-	prefix := fmt.Sprintf("dev:ns:%s:repo:%s:dev:*", namespace, repo)
-	iter := rdb.Scan(c.Request.Context(), 0, prefix, 0).Iterator()
-	for iter.Next(c.Request.Context()) {
-		key := iter.Val()
-		parts := strings.Split(key, ":")
-		// dev:ns:NAMESPACE:repo:REPO:dev:NAME
-		// 0   1  2         3    4    5   6
-		if len(parts) != 7 {
-			continue
-		}
-		name := parts[6]
-
-		data, err := rdb.HGetAll(c.Request.Context(), key).Result()
-		if err != nil {
-			log.Printf("Failed to get DevSandbox %s from Redis for repo %s: %v", name, repo, err)
-			continue
-		}
-
-		sandbox := DevSandbox{
-			Name: name,
-		}
-		if val, ok := data["sandbox"]; ok {
-			sandbox.Sandbox = val
-		}
-		if val, ok := data["sandboxReplica"]; ok {
-			sandbox.SandboxReplica = val
-		}
-		if val, ok := data["branchURL"]; ok {
-			sandbox.BranchURL = val
-		}
-		if val, ok := data["branch"]; ok {
-			sandbox.Branch = val
-		}
-
-		sandboxes = append(sandboxes, sandbox)
-	}
-	if err := iter.Err(); err != nil {
-		log.Printf("Error during Redis SCAN for dev sandboxes: %v", err)
-	}
-
-	c.JSON(http.StatusOK, sandboxes)
-}
-
-func fetchAndPopulateDevSandboxes(ctx context.Context, namespace, repo string) {
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "devsandboxes",
-	}
-	list, err := k8sClient.Resource(gvr).Namespace(namespace).List(context.Background(),
-		v1.ListOptions{
-			LabelSelector: fmt.Sprintf("review.gemini.google.com/repowatch=%s", repo),
-		})
-	if err != nil {
-		log.Printf("Failed to list DevSandbox CRs: %v.", err)
-		return
-	}
-
-	log.Printf("Populating DevSandboxes: Found %d devsandboxes for Repo: %s", len(list.Items), repo)
-	for _, item := range list.Items {
-		replicas, found, err := unstructured.NestedInt64(item.Object, "spec", "replicas")
-		if err != nil || !found {
-			log.Printf("Replicas (.spec.replicas) not found in DevSandbox %s", item.GetName())
-			continue
-		}
-
-		branch, found, err := unstructured.NestedString(item.Object, "spec", "destination", "branch")
-		if err != nil || !found {
-			log.Printf("Branch (.spec.destination.branch) not found in DevSandbox %s", item.GetName())
-			branch = "nobranch" // fallback
-		}
-
-		cloneURL, found, err := unstructured.NestedString(item.Object, "spec", "source", "cloneURL")
-		if err != nil || !found {
-			log.Printf("cloneURL (.spec.source.cloneURL) not found in DevSandbox %s", item.GetName())
-			cloneURL = "https://github.com/noorg/norepo.git"
-		}
-		repoParts := strings.Split(strings.TrimSuffix(cloneURL, ".git"), "/")
-		if len(repoParts) >= 2 {
-			repoName := repoParts[len(repoParts)-1]
-			owner := repoParts[len(repoParts)-2]
-			// Construct branch URL: https://github.com/OWNER/REPO/tree/BRANCH
-			branchURL := fmt.Sprintf("https://github.com/%s/%s/tree/%s", owner, repoName, branch)
-
-			// Store in Redis
-			// Use sandbox name as the identifier for now or the branch name?
-			// The UI card key is `sandbox.name`. If we use branch name, it must be unique per repo.
-			// Let's use the DevSandbox name as the key in Redis to match deletion logic.
-
-			key := fmt.Sprintf("dev:ns:%s:repo:%s:dev:%s", namespace, repo, item.GetName())
-			if err := rdb.HSet(ctx, key,
-				"sandbox", item.GetName(),
-				"branch", branch,
-				"branchURL", branchURL,
-				"sandboxReplica", fmt.Sprintf("%d", replicas),
-			).Err(); err != nil {
-				log.Printf("Failed to cache DevSandbox %s: %v", item.GetName(), err)
-			}
-		}
-	}
-}
-
-func deleteDevSandbox(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	name := c.Param("name") // This is the sandbox name
-	ctx := c.Request.Context()
-
-	if err := scaledownDevSandboxHelper(ctx, namespace, name); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to delete dev sandbox", "details": err.Error()})
-		return
-	}
-
-	key := fmt.Sprintf("dev:ns:%s:repo:%s:dev:%s", namespace, repo, name)
-	if err := rdb.Del(c.Request.Context(), key).Err(); err != nil {
-		log.Printf("Failed to DEL DevSandbox data from Redis: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to DEL DevSandbox data from Redis"})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func scaleUpDevSandbox(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	name := c.Param("name")
-
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "devsandboxes",
-	}
-
-	// Get the existing resource to find max replicas? Or just set to 1.
-	// Usually 1.
-
-	sandbox := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "custom.agents.x-k8s.io/v1alpha1",
-			"kind":       "DevSandbox",
-			"metadata": map[string]interface{}{
-				"name":      name,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"replicas": int64(1),
-			},
-		},
-	}
-
-	_, err := k8sClient.Resource(gvr).Namespace(namespace).Apply(c.Request.Context(), name,
-		sandbox, v1.ApplyOptions{FieldManager: "review-ui", Force: true})
-	if err != nil {
-		log.Printf("Failed to scale up dev sandbox %s: %v", name, err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scale up dev sandbox"})
-		return
-	}
-
-	c.Status(http.StatusOK)
-}
-
-func scaleDownDevSandbox(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	name := c.Param("name")
-	if err := scaledownDevSandboxHelper(c.Request.Context(), namespace, name); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scale down dev sandbox", "details": err.Error()})
-		return
-	}
-	c.Status(http.StatusOK)
-}
-
-func scaledownDevSandboxHelper(ctx context.Context, namespace, name string) error {
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "devsandboxes",
-	}
-	log.Printf("Scaling down dev sandbox %s", name)
-
-	sandbox := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "custom.agents.x-k8s.io/v1alpha1",
-			"kind":       "DevSandbox",
-			"metadata": map[string]interface{}{
-				"name":      name,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"replicas": int64(0),
-			},
-		},
-	}
-
-	_, err := k8sClient.Resource(gvr).Namespace(namespace).Apply(ctx, name,
-		sandbox, v1.ApplyOptions{FieldManager: "review-ui", Force: true})
-	if err != nil {
-		return fmt.Errorf("failed to scaledown dev sandbox: %w", err)
-	}
-	return nil
-}
-
-func proxy(c *gin.Context) {
-	proxyURL := c.Query("url")
-	if proxyURL == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "url query parameter is required"})
-		return
-	}
-
-	// validate the URL begins with  https://github.com/ or https://raw.githubusercontent.com/
-	if !strings.HasPrefix(proxyURL, "https://github.com/") && !strings.HasPrefix(proxyURL, "https://raw.githubusercontent.com/") {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "url must begin with https://github.com/ or https://raw.githubusercontent.com/"})
-		return
-	}
-
-	resp, err := http.Get(proxyURL)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to fetch url: %v", err)})
-		return
-	}
-	defer resp.Body.Close()
-
-	body, err := io.ReadAll(resp.Body)
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to read response body: %v", err)})
-		return
-	}
-
-	c.String(resp.StatusCode, string(body))
-}
-
-// TODO this is k8s specific. we need to generalize it later.
-var allowedLabelPrefixes = []string{"area/", "kind/", "priority/", "sig/", "type/"}
-
-func getSuggestedLabels(ctx context.Context, client *github.Client, owner, repo string) ([][]string, error) {
-	query := fmt.Sprintf("repo:%s/%s involves:@me is:pr", owner, repo)
-	opts := &github.SearchOptions{
-		Sort:        "updated",
-		Order:       "desc",
-		ListOptions: github.ListOptions{PerPage: 100},
-	}
-
-	result, _, err := client.Search.Issues(ctx, query, opts)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, issue := range result.Issues {
-		log.Printf("Found issue #%d with labels:", *issue.Number)
-		for _, label := range issue.Labels {
-			if label.Name != nil {
-				log.Printf(" - %s", *label.Name)
-			}
-		}
-	}
-
-	// Count the label ocurrences
-	issueLabels := [][]string{}
-	unlabelledCount := 0
-	for i, issue := range result.Issues {
-		issueLabels = append(issueLabels, []string{})
-		if len(issue.Labels) == 0 {
-			unlabelledCount++
-		}
-		for _, label := range issue.Labels {
-			if label.Name != nil {
-				// Only consider labels with allowed prefixes
-				allowed := false
-				for _, prefix := range allowedLabelPrefixes {
-					if strings.HasPrefix(*label.Name, prefix) {
-						allowed = true
-						break
-					}
-				}
-				if !allowed {
-					continue
-				}
-				issueLabels[i] = append(issueLabels[i], *label.Name)
-			}
-		}
-	}
-
-	labels, count := findMostCommonCoOccurringLabels(issueLabels)
-
-	// If the most common co-occurring labels appear in less than all unlabelled PRs, return no suggestions
-	if count < unlabelledCount {
-		return [][]string{}, nil
-	}
-
-	return labels, nil
-}
-
-// The size of the co-occurring set we want to find (couplets)
-const TargetSize = 2
-const TopN = 5
-
-// combinations recursively generates all combinations of size k from the list of labels.
-func combinations(labels []string, k int) [][]string {
-	var result [][]string
-
-	var generate func(start int, currentCombination []string)
-	generate = func(start int, currentCombination []string) {
-		// Base case: combination is complete
-		if len(currentCombination) == k {
-			// Create a copy to append to the result
-			combinationCopy := make([]string, k)
-			copy(combinationCopy, currentCombination)
-			result = append(result, combinationCopy)
-			return
-		}
-
-		// Recursive step
-		for i := start; i < len(labels); i++ {
-			// Add the current label and recurse
-			generate(i+1, append(currentCombination, labels[i]))
-		}
-	}
-
-	generate(0, []string{})
-	return result
-}
-
-// findMostCommonCoOccurringLabels processes the data to find the triplet with the highest support.
-func findMostCommonCoOccurringLabels(itemLabels [][]string) ([][]string, int) {
-	// 1. Initialize a map to store the counts (support) of each label set
-	counts := make(map[string]int)
-
-	// 2. Iterate over all items
-	for _, labels := range itemLabels {
-		// Skip items that don't have enough labels
-		if len(labels) < TargetSize {
-			continue
-		}
-
-		// It is crucial to sort the labels *before* generating combinations
-		// to ensure that combinations like (a, b, c) and (c, b, a) are treated
-		// as the same set, ensuring consistent counting.
-		sort.Strings(labels)
-
-		// 3. Generate all combinations of size TargetSize
-		combos := combinations(labels, TargetSize)
-
-		// 4. Count the support for each combination
-		for _, combo := range combos {
-			// Create a unique string key for the map from the sorted combination
-			// This key represents the unique itemset {L1, L2, L3}
-			key := strings.Join(combo, "||")
-			counts[key]++
-		}
-	}
-
-	type labelCount struct {
-		labels []string
-		count  int
-	}
-
-	var allCounts []labelCount
-	for key, count := range counts {
-		allCounts = append(allCounts, labelCount{
-			labels: strings.Split(key, "||"),
-			count:  count,
-		})
-	}
-
-	// Sort by count desc, then by label key asc
-	sort.Slice(allCounts, func(i, j int) bool {
-		if allCounts[i].count != allCounts[j].count {
-			return allCounts[i].count > allCounts[j].count
-		}
-		return strings.Join(allCounts[i].labels, "||") < strings.Join(allCounts[j].labels, "||")
-	})
-
-	var result [][]string
-	maxCount := 0
-	if len(allCounts) > 0 {
-		maxCount = allCounts[0].count
-	}
-
-	for i := 0; i < len(allCounts) && i < TopN; i++ {
-		result = append(result, allCounts[i].labels)
-	}
-
-	return result, maxCount
-}
-
-func scaleUpPR(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	prID := c.Param("id")
-	ctx := c.Request.Context()
-
-	if err := scaleupSandbox(ctx, namespace, repo, prID); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scale up sandbox", "details": err.Error()})
-		return
-	}
-	c.Status(http.StatusOK)
-}
-
-func scaleDownPR(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	prID := c.Param("id")
-	ctx := c.Request.Context()
-
-	if err := scaledownSandbox(ctx, namespace, repo, prID); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scale down sandbox", "details": err.Error()})
-		return
-	}
-	c.Status(http.StatusOK)
-}
-
-func scaleUpIssue(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	issueID := c.Param("issue_id")
-	handler := c.Param("handler")
-	ctx := c.Request.Context()
-
-	if err := scaleupIssueSandbox(ctx, namespace, repo, issueID, handler); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scale up issue sandbox", "details": err.Error()})
-		return
-	}
-	c.Status(http.StatusOK)
-}
-
-func scaleDownIssue(c *gin.Context) {
-	namespace := c.MustGet(userKey).(string)
-	repo := c.Param("repo")
-	issueID := c.Param("issue_id")
-	handler := c.Param("handler")
-	ctx := c.Request.Context()
-
-	if err := scaledownIssueSandbox(ctx, namespace, repo, issueID, handler); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to scale down issue sandbox", "details": err.Error()})
-		return
-	}
-	c.Status(http.StatusOK)
-}
-
-func scaleupSandbox(ctx context.Context, namespace, repo, prID string) error {
-	prKey := fmt.Sprintf("pr:ns:%s:repo:%s:pr:%s", namespace, repo, prID)
-	sandboxName, err := rdb.HGet(ctx, prKey, "sandbox").Result()
-	if err != nil && err != redis.Nil {
-		return fmt.Errorf("failed to get sandbox name from Redis: %w", err)
-	}
-	if sandboxName == "" {
-		sandboxName = fmt.Sprintf("%s-pr-%s", repo, prID)
-	}
-
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "reviewsandboxes",
-	}
-	log.Printf("Scaling up sandbox %s", sandboxName)
-
-	sandbox := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "custom.agents.x-k8s.io/v1alpha1",
-			"kind":       "ReviewSandbox",
-			"metadata": map[string]interface{}{
-				"name":      sandboxName,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"replicas": int64(1),
-			},
-		},
-	}
-
-	_, err = k8sClient.Resource(gvr).Namespace(namespace).Apply(ctx, sandboxName,
-		sandbox, v1.ApplyOptions{FieldManager: "review-ui", Force: true})
-	if err != nil {
-		return fmt.Errorf("failed to scale up sandbox: %w", err)
-	}
-	return nil
-}
-
-func scaleupIssueSandbox(ctx context.Context, namespace, repo, issueID, handler string) error {
-	sandboxName := fmt.Sprintf("%s-issue-%s-%s", repo, issueID, handler)
-
-	gvr := schema.GroupVersionResource{
-		Group:    "custom.agents.x-k8s.io",
-		Version:  "v1alpha1",
-		Resource: "issuesandboxes",
-	}
-	log.Printf("Scaling up issue sandbox %s", sandboxName)
-
-	sandbox := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "custom.agents.x-k8s.io/v1alpha1",
-			"kind":       "IssueSandbox",
-			"metadata": map[string]interface{}{
-				"name":      sandboxName,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"replicas": int64(1),
-			},
-		},
-	}
-
-	_, err := k8sClient.Resource(gvr).Namespace(namespace).Apply(ctx, sandboxName,
-		sandbox, v1.ApplyOptions{FieldManager: "review-ui", Force: true})
-	if err != nil {
-		return fmt.Errorf("failed to scale up issue sandbox: %w", err)
-	}
-	return nil
 }
