+++ repo-agent/review-ui/review-api/pkg/api/fix_yaml_integers_test.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"reflect"
