+++ repo-agent/review-ui/review-api/pkg/api/handlers_common.go
@@ -1,0 +1,44 @@
+package api
+
+import (
+	"fmt"
+	"io"
+	"net/http"
+	"strings"
+
+	"github.com/gin-gonic/gin"
+)
+
+// --- Health Check ---
+func (s *Server) healthCheckOk(c *gin.Context) {
+	c.String(http.StatusOK, "OK")
+}
+
+func (s *Server) proxy(c *gin.Context) {
+	proxyURL := c.Query("url")
+	if proxyURL == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "url query parameter is required"})
+		return
+	}
+
+	// validate the URL begins with  https://github.com/ or https://raw.githubusercontent.com/
+	if !strings.HasPrefix(proxyURL, "https://github.com/") && !strings.HasPrefix(proxyURL, "https://raw.githubusercontent.com/") {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "url must begin with https://github.com/ or https://raw.githubusercontent.com/"})
+		return
+	}
+
+	resp, err := http.Get(proxyURL)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to fetch url: %v", err)})
+		return
+	}
+	defer resp.Body.Close()
+
+	body, err := io.ReadAll(resp.Body)
+	if err != nil {
+		c.JSON(http.StatusInternalServerError, gin.H{"error": fmt.Sprintf("Failed to read response body: %v", err)})
+		return
+	}
+
+	c.String(resp.StatusCode, string(body))
+}
