+++ repo-agent/review-ui/review-api/pkg/api/suggestions_test.go
@@ -1,4 +1,4 @@
-package main
+package api
 
 import (
 	"context"
